requirements.txt
typeguard
synology-api
PyYAML